schema/requirements.txt
# Python dependencies for Catty Thesis Validators
# Install with: pip install -r requirements.txt

pyyaml>=6.0
rdflib>=7.0.0
pyshacl>=0.20.0
requests>=2.28.0
jsonschema>=4.17.0
